crud(postgres+redis)/requirements.txt
Flask==2.0.1
psycopg2-binary==2.9.3
redis==4.3.4
Werkzeug==2.0.3
python-dotenv==0.19.2